Big update! +Highlighting +Possession +Sprite-switching
Currently working possession that allows the player to see the npc that will be possessed via a highlight. Once "e" is pressed, you possess (and teleport to) the npc (with sprite switched). The npc you possessed is now disabled and will be teleported to you once you quit possession by pressing "q"

A handy way to memorize these controls:
e -> Enter the npc's body
q -> Quit possessing the npc