👣 Fixed movement and interaction bugs
The player was able to initiate minigames and walk around in the last few seconds after winning or losing the minigame, before possession happened. This caused the player to initiate minigames unrightfully or walk up to a different NPC to possess. These issues have been fixed by fixing the timing in the PossessionController. The spam minigame has changed to the "x" button again and its files have been moved to a more apt location.